requirements.txt
discord.py
aiohttp
Flask
pytz
pytest
flake8